+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 3. [](#bugfix)
    * Fixed CLI `--env` and `--lang` options having no effect if they aren't added before all the other options
    * Fixed scaled image medium filename when using non-existing retina file
+   * Fixed an issue with JS `imports` and pipelining Assets
 
 # v1.7.26.1
 ## 01/04/2022